+++ databricks/sdk/errors/__init__.py
@@ -1,5 +1,6 @@
-from .base import DatabricksError, ErrorDetail, _get_api_error
+from .base import DatabricksError, ErrorDetail
 from .mapper import error_mapper
+from .parser import _get_api_error
 from .platform import *
 from .private_link import PrivateLinkValidationError
 from .sdk import *
